Move TableOwnership to separate module
